3-print_all.c

+++ 0x10-variadic_functions/3-print_all.c
@@ -1,13 +1,13 @@
 #include "variadic_functions.h"
-#include <stdio.h>
+#include <stdarg.h>
 #include <stdlib.h>
+#include <stdio.h>
 
 /**
- * print_all - prints anything
- * @format: a list of arguments to pass to the function
- * Return: NIl
+ * print_all - A function prints anyting.
+ * @format: A list of type of arguments passed to the function.
+ * Return: Nothing
  */
-
 void print_all(const char * const format, ...)
 {
 	va_list ap;
@@ -31,7 +31,7 @@ void print_all(const char * const format, ...)
 				printf("%d", va_arg(ap, int));
 				break;
 			case 'f':
-				printf("%f", (float) va_arg(ap, int));
+				printf("%f", (float) va_arg(ap, double));
 				break;
 			case 's':
 				temp = va_arg(ap, char*);
@@ -43,8 +43,8 @@ void print_all(const char * const format, ...)
 				printf("(nil)");
 				break;
 		}
-		if ((format[i] == 'c' || format[i] == 'i' || format[i] == 'f'
-					|| format[i] == 's') &&format[(i + 1)] != '\0')
+		if ((format[i] == 'c' || format[i] == 'i' || format[i] == 'f' ||
+					format[i] == 's') && format[(i + 1)] != '\0')
 			printf(", ");
 		i++;
 	}
